requirements.txt
flask
flask-cors
werkzeug
PyMuPDF
langchain
sentence-transformers
chromadb
google-generativeai
python-dotenv
fastapi
uvicorn
jinja2
python-multipart